CXP-1236: Merged Get category tree roots endpoint into GetCategories endpoint
