+++ components/catalogs/back/src/Infrastructure/Symfony/Resources/config/routing/internal.yml
@@ -51,11 +51,6 @@ akeneo_catalogs_internal_get_locales:
     controller: Akeneo\Catalogs\Infrastructure\Controller\Internal\GetLocalesAction
     methods: [ GET ]
 
-akeneo_catalogs_internal_get_category_tree_roots:
-    path: '/rest/catalogs/categories/tree-roots'
-    controller: Akeneo\Catalogs\Infrastructure\Controller\Internal\GetCategoryTreeRootsAction
-    methods: [ GET ]
-
 akeneo_catalogs_internal_get_category_children:
     path: '/rest/catalogs/categories/{categoryId}/children'
     controller: Akeneo\Catalogs\Infrastructure\Controller\Internal\GetCategoryChildrenAction
