+++ components/catalogs/back/tests/Unit/Infrastructure/Controller/Internal/GetCategoriesActionTest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Akeneo\Catalogs\Test\Unit\Infrastructure\Controller\Internal;
 
 use Akeneo\Catalogs\Application\Persistence\GetCategoriesByCodeQueryInterface;
+use Akeneo\Catalogs\Application\Persistence\GetCategoryTreeRootsQueryInterface;
 use Akeneo\Catalogs\Infrastructure\Controller\Internal\GetCategoriesAction;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\HttpFoundation\JsonResponse;
@@ -15,16 +16,21 @@ use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 /**
  * @copyright 2022 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ *
+ * @covers \Akeneo\Catalogs\Infrastructure\Controller\Internal\GetCategoriesAction
  */
 class GetCategoriesActionTest extends TestCase
 {
     private ?GetCategoriesAction $getCategoriesAction;
+    private ?GetCategoryTreeRootsQueryInterface $getCategoryTreeRootsQuery;
     private ?GetCategoriesByCodeQueryInterface $getCategoriesByCodeQuery;
 
     protected function setUp(): void
     {
+        $this->getCategoryTreeRootsQuery = $this->createMock(GetCategoryTreeRootsQueryInterface::class);
         $this->getCategoriesByCodeQuery = $this->createMock(GetCategoriesByCodeQueryInterface::class);
         $this->getCategoriesAction = new GetCategoriesAction(
+            $this->getCategoryTreeRootsQuery,
             $this->getCategoriesByCodeQuery,
         );
     }
@@ -34,18 +40,41 @@ class GetCategoriesActionTest extends TestCase
         self::assertInstanceOf(RedirectResponse::class, ($this->getCategoriesAction)(new Request()));
     }
 
-    public function testItAnswersABadRequestIfTheQueryIsInvalid(): void
+    /**
+     * @dataProvider queryWillThrowDataProvider
+     */
+    public function testItAnswersABadRequestIfTheQueryIsInvalid(array $queryPayload): void
     {
         $this->expectException(BadRequestHttpException::class);
 
         ($this->getCategoriesAction)(
             new Request(
-                query: ['codes' => 123],
+                query: $queryPayload,
                 server: [
                     'HTTP_X-Requested-With' => 'XMLHttpRequest',
                 ],
             )
         );
+    }
+
+    public function queryWillThrowDataProvider(): array
+    {
+        return [
+            'must specify either codes or is_root' => [
+                [],
+            ],
+            'codes must be a string' => [
+                [
+                    'codes' => 456,
+                ],
+            ],
+            'codes and is_root are mutually exclusive' => [
+                [
+                    'codes' => 'a,b,c',
+                    'is_root' => true,
+                ],
+            ],
+        ];
     }
 
     public function testItReturnsCategoriesFromTheQuery(): void
@@ -58,6 +87,28 @@ class GetCategoriesActionTest extends TestCase
         $response = ($this->getCategoriesAction)(
             new Request(
                 query: ['codes' => 'codeA,codeB,codeC'],
+                server: [
+                    'HTTP_X-Requested-With' => 'XMLHttpRequest',
+                ],
+            )
+        );
+
+        self::assertInstanceOf(JsonResponse::class, $response);
+        self::assertJsonStringEqualsJsonString(
+            \json_encode(['categoryA', 'categoryB', 'categoryC'], JSON_THROW_ON_ERROR),
+            $response->getContent()
+        );
+    }
+
+    public function testItReturnsRootCategoriesFromTheQuery(): void
+    {
+        $this->getCategoryTreeRootsQuery
+            ->method('execute')
+            ->willReturn(['categoryA', 'categoryB', 'categoryC']);
+
+        $response = ($this->getCategoriesAction)(
+            new Request(
+                query: ['is_root' => true],
                 server: [
                     'HTTP_X-Requested-With' => 'XMLHttpRequest',
                 ],
